refactor: update group model package
- break out access/permission-related methods inside `group.go` into `access.go`
- add `Capabilities` struct and associated `GetCapabilities` computation method
